Use JSX for modal
